+++ lib/recurly.js
@@ -18,8 +18,9 @@ import {Bus} from './recurly/bus';
 import {Fraud} from './recurly/fraud';
 import {HostedFields, FIELD_TYPES} from './recurly/hosted-fields';
 import {Message} from './recurly/message';
+import CheckoutPricing from './recurly/pricing/checkout';
+import SubscriptionPricing from './recurly/pricing/subscription';
 import {uuid} from './util/uuid';
-import Pricing from './recurly/pricing';
 
 import giftcard from './recurly/giftcard';
 
@@ -84,7 +85,7 @@ const defaults = {
  * @public
  */
 
-export class Recurly extends Emitter {
+export default class Recurly extends Emitter {
   constructor (options) {
     super();
     this.id = uuid();
@@ -97,6 +98,8 @@ export class Recurly extends Emitter {
       token: bankAccount.token.bind(this),
       bankInfo: bankAccount.bankInfo.bind(this)
     };
+    this.Pricing.Checkout = CheckoutPricing;
+    this.Pricing.Subscription = SubscriptionPricing;
   }
 
   ready (done) {
@@ -366,8 +369,13 @@ export class Recurly extends Emitter {
     });
   }
 
+  /**
+   * Old method of instantiating single-subscription Pricing
+   *
+   * @deprecated
+   */
   Pricing () {
-    return new Pricing(this);
+    return new SubscriptionPricing(this);
   }
 }
