+++ lib/recurly/apple-pay.js
@@ -1,6 +1,6 @@
 import Emitter from 'component-emitter';
 import errors from '../errors';
-import {Pricing} from './pricing';
+import {SubscriptionPricing} from './pricing/subscription';
 import {normalize} from '../util/normalize';
 import {FIELDS} from './token';
 
@@ -212,6 +212,7 @@ class ApplePay extends Emitter {
    * @private
    */
   onPricingChange (price) {
+    // TODO: Update this to react to CheckoutPricing or SubscriptionPricing
     if (!price) return;
     // Reset line items
     this.config.lineItems = [];
@@ -222,7 +223,7 @@ class ApplePay extends Emitter {
    * Validates the merchant session
    *
    * @param  {Event} event ApplePayValidateMerchantEvent
-   * @return {[type]}       [description]
+   * @return {[type]} [description]
    * @private
    */
   onValidateMerchant (event) {
