+++ lib/recurly/pricing/subscription/attachment.js
@@ -1,6 +1,6 @@
 import Emitter from 'component-emitter';
 import find from 'component-find';
-import dom from '../../util/dom';
+import dom from '../../../util/dom';
 
 const debug = require('debug')('recurly:pricing:attachment');
