+++ package.json
@@ -18,6 +18,7 @@
     "json-component": "0.0.1",
     "jsonp": "0.0.4",
     "load-script": "^1.0.0",
+    "lodash.isempty": "^4.4.0",
     "lodash.map": "^4.6.0",
     "lodash.merge": "^4.6.0",
     "lodash.omit": "^4.5.0",
